Implement Goal Detection Driver
Goal detection uses external interrupts to mark if a player scores by setting the correct value in a ScoreState enum.

This enum can be checked while the game is running to see if a player scored and can switch states accordingly.

Since goal detection does not switch states, there is no issue with multiple external interrupts on the same goal being raised.